Fix enum get/set to match backing type: int→int, uint→uint, long→nint, ulong→nuint

+++ Metal.NET/Metal/MTL4PrimitiveAccelerationStructureDescriptor.cs
@@ -15,7 +15,7 @@ public class MTL4PrimitiveAccelerationStructureDescriptor(nint nativePtr) : MTL4
     public MTLMotionBorderMode MotionEndBorderMode
     {
         get => (MTLMotionBorderMode)ObjectiveCRuntime.MsgSendUInt(NativePtr, MTL4PrimitiveAccelerationStructureDescriptorBindings.MotionEndBorderMode);
-        set => ObjectiveCRuntime.MsgSend(NativePtr, MTL4PrimitiveAccelerationStructureDescriptorBindings.SetMotionEndBorderMode, (nuint)value);
+        set => ObjectiveCRuntime.MsgSend(NativePtr, MTL4PrimitiveAccelerationStructureDescriptorBindings.SetMotionEndBorderMode, (uint)value);
     }
 
     public float MotionEndTime
@@ -33,7 +33,7 @@ public class MTL4PrimitiveAccelerationStructureDescriptor(nint nativePtr) : MTL4
     public MTLMotionBorderMode MotionStartBorderMode
     {
         get => (MTLMotionBorderMode)ObjectiveCRuntime.MsgSendUInt(NativePtr, MTL4PrimitiveAccelerationStructureDescriptorBindings.MotionStartBorderMode);
-        set => ObjectiveCRuntime.MsgSend(NativePtr, MTL4PrimitiveAccelerationStructureDescriptorBindings.SetMotionStartBorderMode, (nuint)value);
+        set => ObjectiveCRuntime.MsgSend(NativePtr, MTL4PrimitiveAccelerationStructureDescriptorBindings.SetMotionStartBorderMode, (uint)value);
     }
 
     public float MotionStartTime
